Add state type coercion
Sometimes a state is persisted and later an attempt is made to restore through initial state. If the schema has changed the state may become incompatible. This commit adds the ability to coerce state data to the right state shape, by setting missing or incorrect data to default and ignoring extra data. This is a possibly lossy process. New modifier "coerce" added to allow the developer to specify how the data should be coerced / upgraded.